# frontend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ChatRequest" : {
        "properties" : {
          "criticality" : {
            "type" : "string"
          },
          "deadlineMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "maxTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "measureUncertainty" : {
            "type" : "boolean"
          },
          "messages" : {
            "items" : {
              "$ref" : "#/components/schemas/Message"
            },
            "type" : "array"
          },
          "model" : {
            "type" : "string"
          },
          "requireVision" : {
            "type" : "boolean"
          },
          "responseFormat" : {
            "$ref" : "#/components/schemas/ResponseFormatRef"
          },
          "routingMode" : {
            "type" : "string"
          },
          "stream" : {
            "type" : "boolean"
          },
          "temperature" : {
            "format" : "double",
            "type" : "number"
          },
          "toolChoice" : {
            "type" : "string"
          },
          "tools" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "Event" : {
        "properties" : {
          "createdAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "detail" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "model" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GatewayRequestLogEntity" : {
        "properties" : {
          "chosenModel" : {
            "type" : "string"
          },
          "chosenProvider" : {
            "type" : "string"
          },
          "complexity" : {
            "format" : "double",
            "type" : "number"
          },
          "costUsd" : {
            "format" : "double",
            "type" : "number"
          },
          "createdAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "developerId" : {
            "type" : "string"
          },
          "failoverCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "latencyMs" : {
            "format" : "int64",
            "type" : "integer"
          },
          "requestedModel" : {
            "type" : "string"
          },
          "routingReason" : {
            "type" : "string"
          },
          "success" : {
            "type" : "boolean"
          },
          "tokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "traceId" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ImageRef" : {
        "properties" : {
          "detail" : {
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "InboundResult" : {
        "properties" : {
          "blocked" : {
            "type" : "boolean"
          },
          "injectionHits" : {
            "items" : {
              "type" : "string"
            },
            "type" : "array"
          },
          "injectionScore" : {
            "format" : "double",
            "type" : "number"
          },
          "redactions" : {
            "items" : {
              "$ref" : "#/components/schemas/Match"
            },
            "type" : "array"
          },
          "sanitized" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Match" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "count" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "Message" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "images" : {
            "items" : {
              "$ref" : "#/components/schemas/ImageRef"
            },
            "type" : "array"
          },
          "role" : {
            "type" : "string"
          },
          "toolCallId" : {
            "type" : "string"
          },
          "toolCalls" : {
            "items" : {
              "$ref" : "#/components/schemas/ToolCallRef"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
      "ModelEntity" : {
        "properties" : {
          "capabilitiesJson" : {
            "type" : "string"
          },
          "chat" : {
            "type" : "boolean"
          },
          "contextWindow" : {
            "format" : "int32",
            "type" : "integer"
          },
          "description" : {
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "firstSeenAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "kind" : {
            "type" : "string"
          },
          "lastCheckedAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "limitsJson" : {
            "type" : "string"
          },
          "maxOutputTokens" : {
            "format" : "int32",
            "type" : "integer"
          },
          "missCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "missingSince" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "modelName" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "preview" : {
            "type" : "boolean"
          },
          "pricingJson" : {
            "type" : "string"
          },
          "probeOutcome" : {
            "type" : "string"
          },
          "probedAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "provider" : {
            "type" : "string"
          },
          "providerCreatedAt" : {
            "format" : "int64",
            "type" : "integer"
          },
          "replacedBy" : {
            "type" : "string"
          },
          "retiredAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "source" : {
            "type" : "string"
          },
          "status" : {
            "enum" : [ "DISCOVERED", "TESTING", "ACTIVE", "DEPRECATED", "REMOVED", "UNAVAILABLE" ],
            "type" : "string"
          },
          "statusReason" : {
            "type" : "string"
          },
          "verifiedAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          }
        },
        "type" : "object"
      },
...
      "PinRequest" : {
        "properties" : {
          "model" : {
            "type" : "string"
          },
          "provider" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ResponseFormatRef" : {
        "properties" : {
          "schema" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          },
          "type" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Run" : {
        "properties" : {
          "finishedAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "listCalls" : {
            "format" : "int32",
            "type" : "integer"
          },
          "outcome" : {
            "type" : "string"
          },
          "probeCalls" : {
            "format" : "int32",
            "type" : "integer"
          },
          "requestedBy" : {
            "type" : "string"
          },
          "startedAt" : {
            "description" : "Epoch seconds, possibly fractional",
            "type" : "number"
          },
          "summary" : {
            "type" : "string"
          },
          "trigger" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TextRequest" : {
        "properties" : {
          "text" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "ToolCallRef" : {
        "properties" : {
          "argumentsJson" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ToolRef" : {
        "properties" : {
          "description" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "parameters" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Durable execution engine for AI agents: workflows, gateway and console APIs.",
    "title" : "Continuum",
    "version" : "1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/gateway/chat" : {
      "post" : {
        "operationId" : "chat",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChatRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "developer-gateway-controller" ]
      }
    },
...
    "/api/gateway/requests" : {
      "get" : {
        "operationId" : "requests_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 100,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GatewayRequestLogEntity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "gateway-stats-controller" ]
      }
    },
...
    "/api/models/events" : {
      "get" : {
        "operationId" : "events_2",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 60,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Event"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "model-catalog-controller" ]
      }
    },
    "/api/models/pin" : {
      "post" : {
        "operationId" : "pin",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PinRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "model-catalog-controller" ]
      }
    },
    "/api/models/runs" : {
      "get" : {
        "operationId" : "runs",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/Run"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "model-catalog-controller" ]
      }
    },
    "/api/models/{id}/status" : {
      "post" : {
        "operationId" : "transition",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "in" : "query",
          "name" : "status",
          "required" : true,
          "schema" : {
            "enum" : [ "DISCOVERED", "TESTING", "ACTIVE", "DEPRECATED", "REMOVED", "UNAVAILABLE" ],
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ModelEntity"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "model-registry-controller" ]
      }
    },
...
    "/api/portal/developer/demo/firewall" : {
      "post" : {
        "operationId" : "firewall_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/TextRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/InboundResult"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "demo-controller" ]
      }
    },
...
    "/api/portal/developer/provenance/requests" : {
      "get" : {
        "operationId" : "requests_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 30,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "type" : "string"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "provenance-controller" ]
      }
    },
...
    "/api/portal/developer/requests" : {
      "get" : {
        "operationId" : "requests",
        "parameters" : [ {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GatewayRequestLogEntity"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "portal-developer-controller" ]
      }
    },
...
    "/api/portal/developer/specialists/catalogue" : {
      "get" : {
        "operationId" : "catalogue",
        "parameters" : [ {
          "in" : "query",
          "name" : "q",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "limit",
          "required" : false,
          "schema" : {
            "default" : 25,
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "specialist-connection-controller" ]
      }
    },
...
    "/api/portal/developer/v8/firewall/{action}" : {
      "post" : {
        "operationId" : "firewall",
        "parameters" : [ {
          "in" : "path",
          "name" : "action",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "additionalProperties" : {
                    "type" : "object"
                  },
                  "type" : "object"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "v-8-controller" ]
      }
    },
...
  }
}